🐛 Fix nil pointer panic when switching to aggregate mode
The node changed callback was trying to access a.client.Cluster.Nodes
after we set a.client = nil in aggregate mode, causing a panic.

Fixed by keeping a.client around in aggregate mode instead of setting
it to nil. The client won't be used (we use a.aggregateManager instead),
but keeping it prevents callbacks from breaking.

This maintains backward compatibility with callbacks set up during
initialization that assume a.client is always available.